🧬 Migrate the source code about subcommand *run* process into component layer in new sub-package *command.run*.
